Implement confirm modal of FinancialRequestDetain in saga

+++ src/store/modules/profile/sagas.js
@@ -40,10 +40,13 @@ const deleteProfile = createApiCallSaga({
 })
 
 const deleteProfileAndRefresh = function*(action) {
-  yield deleteProfile(action)
-  yield getProfiles({
-    type: Types.GET_PROFILES
-  })
+  const confirmed = yield confirm(action.payload.message)
+  if (confirmed) {
+    yield deleteProfile(action)
+    yield getProfiles({
+      type: Types.GET_PROFILES
+    })
+  }
 }
 
 const createProfile = createApiCallSaga({
@@ -61,17 +64,11 @@ const createProfile = createApiCallSaga({
   }
 })
 
-const processDeleteAndRefreshProfile = function*(action) {
-  const confirmed = yield confirm(action.payload.message)
-  if (!confirmed) return
-  yield deleteProfileAndRefresh(action)
-}
-
 export default function* rootSaga() {
   yield takeLatest(Types.GET_PROFILES, getProfiles)
   yield takeLatest(Types.GET_PROFILE_DETAIL, getProfileDetail)
   yield takeLatest(Types.UPDATE_PROFILE, updateProfile)
   yield takeLatest(Types.DELETE_PROFILE, deleteProfile)
-  yield takeLatest(Types.DELETE_PROFILE_AND_REFRESH, processDeleteAndRefreshProfile)
+  yield takeLatest(Types.DELETE_PROFILE_AND_REFRESH, deleteProfileAndRefresh)
   yield takeLatest(Types.CREATE_PROFILE, createProfile)
 }
